### 2018-10-05 #998 @whedon
[PRE REVIEW]: SeqTools: A python package for easy transformation, combination and evaluation of large datasets
**Submitting author:** @nlgranger (<a href="http://orcid.org/0000-0003-2943-9888">Nicolas Granger</a>)
**Repository:** <a href="https://github.com/nlgranger/SeqTools" target ="_blank">https://github.com/nlgranger/SeqTools</a>
**Version:** v0.9.0
