### 2018-10-08 #998 @arfon
@mgymrek @pjotrp - would either if you be willing to edit this submission for JOSS?
