### 2018-10-09 945d2f2
added references to other works, fixed orcid id
