+++ README.rst
@@ -64,9 +64,9 @@ fault-tolerant functions and insightful error reporting. Moreover, internal code
 is kept concise and clear with comments to facilitate error tracing through a
 failing transformation pipeline.
 
-The library should connect nicely to the input pipeline of Machine Learning
-libraries. Nevertheless, this project purposedly keeps a generic interface and
-only requires minimal dependencies in order to facilitate reusability.
+Nevertheless, this project purposedly keeps a generic interface and only
+requires minimal dependencies in order to facilitate reusability beyond this
+scope of application.
 
 
 Example
@@ -158,10 +158,22 @@ The documentation is hosted at https://seqtools-doc.readthedocs.io
 Related libraries
 -----------------
 
-These libaries provide comparable functionalities, but for iterable containers
-only, they should plug easily above a SeqTool based preprocessing pipeline:
+`Joblib <https://joblib.readthedocs.io>`_, proposes low-level functions with
+many optimization settings to optimize pipelined transformations. This library
+notably provides advanced caching mechanisms which are not the primary concern
+of SeqTool. SeqTool uses a simpler container-oriented interface with multiple
+utility functions in order to assist fast prototyping. On-demand evaluation is
+its default behaviour and applies at all layers of a transformation pipeline. In
+particular, parallel evaluation can be inserted in the middle of the
+transformation pipeline and won't block the execution to wait for the
+computation of all elements from the dataset.
 
-- `torchvision.transforms
-  <http://pytorch.org/docs/master/torchvision/transforms.html>`_
-  and `torch.utils.data <http://pytorch.org/docs/master/data.html>`_.
-- `TensorPack <https://github.com/tensorpack/tensorpack>`_
+SeqTools is conceived to connect nicely to the data loading pipeline of Machine
+Learning libraries such as PyTorch's `torch.utils.data
+<http://pytorch.org/docs/master/data.html>`_ and `torchvision.transforms
+<http://pytorch.org/docs/master/torchvision/transforms.html>`_ or Tensorflow's
+`tf.data <https://www.tensorflow.org/guide/datasets>`_. The interface of these
+libraries focuses on `iterators
+<https://docs.python.org/3/library/stdtypes.html#iterator-types>`_ to access
+transformed elements, contary to SeqTools which also provides arbitrary reads
+via indexing.
