### 2018-10-09 #1006 @pjotrp
@nlgranger to expedite the review process do you mind going through above list of check boxes and make sure they can be ticked (you can't tick them). Also check the PDF output carefully. Ping us here when you are done.
