### 2018-10-09 8537e1d
added contribution guidelines
