### 2018-10-09 #1006 @pjotrp
Thanks! You can bump before final publishing.

@Benjamin-Lee you can start review. Please read the reviewer guidelines above
