### 2018-10-23 940a0db
quickfix: hardened stack formatting
