+++ docs/conf.py
@@ -127,61 +127,11 @@ else:
 html_show_sourcelink = False
 
 
-# -- Options for HTMLHelp output ----------------------------------------------
+# -- Options for doctests extension -------------------------------------------
 
-# Output file base name for HTML help builder.
-htmlhelp_basename = 'SeqTools_doc'
-
-
-# -- Options for LaTeX output -------------------------------------------------
-
-latex_elements = {
-    # The paper size ('letterpaper' or 'a4paper').
-    #
-    # 'papersize': 'letterpaper',
-
-    # The font size ('10pt', '11pt' or '12pt').
-    #
-    # 'pointsize': '10pt',
-
-    # Additional stuff for the LaTeX preamble.
-    #
-    # 'preamble': '',
-
-    # Latex figure (float) alignment
-    #
-    # 'figure_align': 'htbp',
-}
-
-# Grouping the document tree into LaTeX files. List of tuples
-# (source start file, target name, title,
-#  author, documentclass [howto, manual, or own class]).
-latex_documents = [
-    (master_doc, 'SeqTools.tex', 'SeqTools Documentation',
-     'Nicolas Granger', 'manual'),
-]
-
-
-# -- Options for manual page output -------------------------------------------
-
-# One entry per manual page. List of tuples
-# (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, 'seqtools', 'SeqTools Documentation',
-     [author], 1)
-]
-
-
-# -- Options for Texinfo output -----------------------------------------------
-
-# Grouping the document tree into Texinfo files. List of tuples
-# (source start file, target name, title, author,
-#  dir menu entry, description, category)
-texinfo_documents = [
-    (master_doc, 'SeqTools', 'SeqTools Documentation',
-     author, 'SeqTools', 'One line description of project.',
-     'Miscellaneous'),
-]
+doctest_global_setup = '''
+import seqtools
+'''
 
 
 # -- Options for Linkcode extension -------------------------------------------

[…]
