### 2018-10-23 ff7f2ae
provide more explicit community guidelines
