### 2018-10-24 d608546
typo, fixes issue #3
