### 2018-10-26 #1006 @nlgranger
Hi, thank you all for the review.
The bibliography items do not have a DOI (one is a white-paper, the other a workshop paper).
The doi is: https://doi.org/10.5281/zenodo.1472094
