### 2018-10-26 #1006 @pjotrp
@arfon we are ready to R&R
