### 2018-10-26 #1006 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
